requirements.txt
# Core dependencies
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# Optional: PDF support
pypdfium2>=4.0.0

# CLI
click>=8.1.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.0.280
mistralai>=1.0.0
